refactor: priceAggregatorChainlink -> fluxAggregator
