+++ packages/inter-protocol/scripts/price-feed-core.js
@@ -57,8 +57,8 @@ export const defaultProposalBuilder = async (
         brandOutRef: brandOut && publishRef(brandOut),
         priceAggregatorRef: publishRef(
           install(
-            '@agoric/inter-protocol/src/price/priceAggregatorChainlink.js',
-            '../bundles/bundle-priceAggregatorChainlink.js',
+            '@agoric/inter-protocol/src/price/fluxAggregator.js',
+            '../bundles/bundle-fluxAggregator.js',
           ),
         ),
       },
